requirements.txt
Streamlit version: 1.38.0
Torch version: 2.4.1+cu118
Transformers version: 4.46.0.dev0
Pillow (PIL) version: 10.4.0